improve project structure and refactor code

+++ aoc10/aoc10.h
@@ -1,11 +1,10 @@
 #ifndef AOC10_H
 #define AOC10_H
 
-#include <glib.h>
-#include "../shared/matrix.h"
+#include <aoc_shared.h>
 
-long long get_total_trailhead_score(char* map);
+gint64 get_total_trailhead_score(char* map);
 
-long long get_total_trailhead_rating(char* map);
+gint64 get_total_trailhead_rating(char* map);
 
 #endif //AOC10_H
