+++ aoc12/aoc12.h
@@ -1,9 +1,9 @@
 #ifndef AOC12_H
 #define AOC12_H
 
-#include "../shared/matrix.h"
-#include "../shared/aocmath.h"
+#include "../shared/aoc_matrix.h"
+#include "../shared/aoc_math.h"
 
-long long get_total_fence_price(char* map, gboolean with_bulk_discount);
+gint64 get_total_fence_price(char* map, gboolean with_bulk_discount);
 
 #endif //AOC12_H
